front_end/requirements.txt
psycopg2-binary~=2.8.6
python-decouple~=3.4
pandas==1.3.2
numpy==1.21.2
sqlalchemy~=1.4.25
beautifulsoup4==4.11.1
requests==2.31.0
python-dotenv==1.0.0
